requirements.txt
streamlit==1.31.0
spacy==3.7.2
pdfminer.six==20231228
python-docx==1.1.0
rapidfuzz==3.6.1
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.4